Move getDomain to BeaconStateAccessors
