+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/common/helpers/BeaconStateAccessors.java
@@ -29,6 +29,7 @@ import tech.pegasys.teku.bls.BLSPublicKey;
 import tech.pegasys.teku.infrastructure.collections.TekuPair;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.config.SpecConfig;
+import tech.pegasys.teku.spec.datastructures.state.Fork;
 import tech.pegasys.teku.spec.datastructures.state.Validator;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconState;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconStateCache;
@@ -295,5 +296,26 @@ public abstract class BeaconStateAccessors {
         "Committee information must be derived from a state no older than the previous epoch. State at slot %s is older than cutoff slot %s",
         state.getSlot(),
         oldestQueryableSlot);
+  }
+
+  public Bytes32 getDomain(BeaconState state, Bytes4 domainType, UInt64 messageEpoch) {
+    UInt64 epoch = (messageEpoch == null) ? getCurrentEpoch(state) : messageEpoch;
+    return getDomain(domainType, epoch, state.getFork(), state.getGenesis_validators_root());
+  }
+
+  public Bytes32 getDomain(BeaconState state, Bytes4 domainType) {
+    return getDomain(state, domainType, null);
+  }
+
+  public Bytes32 getDomain(
+      final Bytes4 domainType,
+      final UInt64 epoch,
+      final Fork fork,
+      final Bytes32 genesisValidatorsRoot) {
+    Bytes4 forkVersion =
+        (epoch.compareTo(fork.getEpoch()) < 0)
+            ? fork.getPrevious_version()
+            : fork.getCurrent_version();
+    return miscHelpers.computeDomain(domainType, forkVersion, genesisValidatorsRoot);
   }
 }
